+++ src/containers/EditListingPage/EditListingWizard/EditListingAvailabilityPanel/EditListingAvailabilityForm.js
@@ -1,0 +1,98 @@
+import React, { Component } from 'react';
+import { bool, func, object, string } from 'prop-types';
+import { compose } from 'redux';
+import classNames from 'classnames';
+import { Form as FinalForm } from 'react-final-form';
+
+// Import configs and util modules
+import { intlShape, injectIntl, FormattedMessage } from '../../../../util/reactIntl';
+import { propTypes } from '../../../../util/types';
+
+// Import shared components
+import { Form, Button } from '../../../../components';
+
+// Import modules from this directory
+import ManageAvailabilityCalendar from './ManageAvailabilityCalendar';
+import css from './EditListingAvailabilityForm.module.css';
+
+export class EditListingAvailabilityFormComponent extends Component {
+  render() {
+    return (
+      <FinalForm
+        {...this.props}
+        render={formRenderProps => {
+          const {
+            className,
+            rootClassName,
+            disabled,
+            ready,
+            handleSubmit,
+            //intl,
+            invalid,
+            pristine,
+            saveActionMsg,
+            updated,
+            updateError,
+            updateInProgress,
+            availability,
+            availabilityPlan,
+            listingId,
+          } = formRenderProps;
+
+          const errorMessage = updateError ? (
+            <p className={css.error}>
+              <FormattedMessage id="EditListingAvailabilityForm.updateFailed" />
+            </p>
+          ) : null;
+
+          const classes = classNames(rootClassName || css.root, className);
+          const submitReady = (updated && pristine) || ready;
+          const submitInProgress = updateInProgress;
+          const submitDisabled = invalid || disabled || submitInProgress;
+
+          return (
+            <Form className={classes} onSubmit={handleSubmit}>
+              {errorMessage}
+              <div className={css.calendarWrapper}>
+                <ManageAvailabilityCalendar
+                  availability={availability}
+                  availabilityPlan={availabilityPlan}
+                  listingId={listingId}
+                />
+              </div>
+
+              <Button
+                className={css.submitButton}
+                type="submit"
+                inProgress={submitInProgress}
+                disabled={submitDisabled}
+                ready={submitReady}
+              >
+                {saveActionMsg}
+              </Button>
+            </Form>
+          );
+        }}
+      />
+    );
+  }
+}
+
+EditListingAvailabilityFormComponent.defaultProps = {
+  updateError: null,
+};
+
+EditListingAvailabilityFormComponent.propTypes = {
+  intl: intlShape.isRequired,
+  onSubmit: func.isRequired,
+  saveActionMsg: string.isRequired,
+  disabled: bool.isRequired,
+  ready: bool.isRequired,
+  updated: bool.isRequired,
+  updateError: propTypes.error,
+  updateInProgress: bool.isRequired,
+  availability: object.isRequired,
+  availabilityPlan: propTypes.availabilityPlan.isRequired,
+};
+
+export default compose(injectIntl)(EditListingAvailabilityFormComponent);